+++ src/collective/cover/tests/test_upgrades.py
@@ -524,8 +524,6 @@ class Upgrade13to14TestCase(UpgradeTestCaseBase):
         self.assertGreaterEqual(int(version), int(self.to_version))
         self.assertEqual(self._how_many_upgrades_to_do(), 4)
 
-    # FIXME: https://github.com/collective/collective.cover/issues/633
-    @unittest.skipIf(IS_PLONE_5, 'Upgrade step not supported under Plone 5')
     def test_register_calendar_tile(self):
         # address also an issue with Setup permission
         title = u'Register calendar tile'